pyproject.toml
[project]
name = "lightweight-web-agent"
version = "0.1.0"
description = "轻量级 Web GUI Agent - 基于 browser-use 的简化版本"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

